page refactoring
