+++ frontend/src/app/Router.tsx
@@ -1,9 +1,9 @@
 import { createBrowserRouter, Navigate } from "react-router";
 import { Layout } from "../widgets/layout/ui/Layout";
-import { Dashboard } from "../pages/dashboard";
-import { FirmwarePage } from "../pages/firmware";
-import { Advertisement } from "../pages/advertisement";
-import { Monitoring } from "../pages/monitoring";
+import { DashboardPage } from "../pages/DashboardPage";
+import { FirmwarePage } from "../pages/FirmwarePage";
+import { AdsPage } from "../pages/AdsPage";
+import { MonitoringPage } from "../pages/MonitoringPage";
 import { FirmwareDetailPage } from "../pages/FirmwareDetailPage";
 
 export const Router = createBrowserRouter([
@@ -13,11 +13,11 @@ export const Router = createBrowserRouter([
     children: [
       // 초기 엔트리 경로를 dashboard로 지정
       { index: true, element: <Navigate to="/dashboard" replace /> },
-      { path: "dashboard", element: <Dashboard /> },
+      { path: "dashboard", element: <DashboardPage /> },
       { path: "firmware", element: <FirmwarePage /> },
       { path: "firmware/:id", element: <FirmwareDetailPage /> },
-      { path: "ads", element: <Advertisement /> },
-      { path: "monitoring", element: <Monitoring /> },
+      { path: "ads", element: <AdsPage /> },
+      { path: "monitoring", element: <MonitoringPage /> },
     ],
   },
 ]);
