+++ frontend/src/pages/AdsPage.tsx
@@ -1,6 +1,6 @@
 import { TitleTile } from "../widgets/layout/ui/TitleTile";
 
-export const Advertisement = () => {
+export const AdsPage = () => {
   return (
     <div>
       <TitleTile title="광고 관리" description="광고 업로드 및 스케줄링" />
