+++ frontend/src/pages/DashboardPage.tsx
@@ -1,6 +1,6 @@
 import { TitleTile } from "../widgets/layout/ui/TitleTile";
 
-export const Dashboard = () => {
+export const DashboardPage = () => {
   return (
     <div>
       <TitleTile title="대시보드" description="디바이스 정보 및 최근 활동" />
