+++ frontend/src/pages/FirmwarePage.tsx
