+++ frontend/src/pages/MonitoringPage.tsx
@@ -1,6 +1,6 @@
 import { TitleTile } from "../widgets/layout/ui/TitleTile";
 
-export const Monitoring = () => {
+export const MonitoringPage = () => {
   return (
     <div>
       <TitleTile title="모니터링" description="기기 상태 모니터링" />
